requirements.txt
asgiref==3.8.1
Django==5.1.5
django-debug-toolbar==5.0.1
django-environ==0.12.0
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
